Move command formatting to transport
