--- nixops/transports/types.py
@@ -1,4 +1,0 @@
-from typing import Union, Iterable
-
-
-Command = Union[str, Iterable[str]]
